Comprehensive code cleanup and refactoring
This commit implements a major cleanup and refactoring effort that removes
unused code, extracts reusable components, and improves code organization.

## Removed Files (1,907 lines removed)
- main_minimal.dart: Alternative main file no longer needed
- simple_flight_list_screen.dart: Duplicate flight list implementation
- flight_track_canvas_screen.dart: Unused track visualization screen
- lib/scripts/: Database utility scripts (fix_database.dart, reset_database.dart)
- tools/: Site data processing utilities no longer needed

## New Shared Components (1,196 lines added)
- lib/presentation/widgets/edit_site_dialog.dart: Extracted 380-line dialog from ManageSitesScreen
- lib/presentation/widgets/site_selection_dialog.dart: Extracted 303-line dialog with search and country grouping
- lib/presentation/widgets/wing_selection_dialog.dart: Extracted 205-line dialog with manufacturer/model search
- lib/utils/date_time_utils.dart: Centralized duration formatting utilities
- lib/utils/ui_utils.dart: Shared error handling and UI utilities

## Refactored Files
- FlightDetailScreen: Reduced from 1,870 to 1,345 lines (525 lines removed)
  - Extracted site and wing selection dialogs to shared widgets
  - Improved code maintainability and reusability
- ManageSitesScreen: Reduced by 380 lines by extracting edit site dialog
- Multiple screens updated to use centralized utilities

## Dependencies
- Removed unused 'provider' package from pubspec.yaml

## Test Fixes
- Fixed async/await issues in site_matching_test.dart
- Tests now compile and run correctly (some API-dependent test failures remain)

## Impact
- Net reduction: ~1,950 lines of code removed
- Improved maintainability through component extraction
- Better code reuse and consistency
- Cleaner project structure
