--- free_flight_log_app/lib/main_minimal.dart
@@ -1,54 +1,0 @@
-import 'package:flutter/material.dart';
-
-void main() {
-  runApp(const MinimalApp());
-}
-
-class MinimalApp extends StatelessWidget {
-  const MinimalApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Free Flight Log',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue),
-        useMaterial3: true,
-      ),
-      home: const MinimalScreen(),
-    );
-  }
-}
-
-class MinimalScreen extends StatelessWidget {
-  const MinimalScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Free Flight Log'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-      ),
-      body: const Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(Icons.flight_takeoff, size: 64),
-            SizedBox(height: 16),
-            Text('Free Flight Log App Running!'),
-            Text('Database and UI ready for development'),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text('App is working!')),
-          );
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
