+++ free_flight_log_app/lib/presentation/screens/flight_detail_screen.dart
@@ -10,6 +10,9 @@ import '../../data/repositories/site_repository.dart';
 import '../../data/repositories/wing_repository.dart';
 import '../widgets/flight_track_widget.dart';
 import '../widgets/flight_statistics_widget.dart';
+import '../widgets/edit_site_dialog.dart';
+import '../widgets/site_selection_dialog.dart';
+import '../widgets/wing_selection_dialog.dart';
 import 'flight_track_screen.dart';
 
 class FlightDetailScreen extends StatefulWidget {
@@ -123,9 +126,9 @@ class _FlightDetailScreenState extends State<FlightDetailScreen> {
       
       if (!mounted) return;
       
-      final result = await showDialog<_SiteSelectionResult>(
+      final result = await showDialog<SiteSelectionResult>(
         context: context,
-        builder: (context) => _SiteSelectionDialog(
+        builder: (context) => SiteSelectionDialog(
           sites: sites,
           currentSite: isLaunchSite ? _launchSite : null,
           title: isLaunchSite ? 'Select Launch Site' : 'Select Landing Site',
@@ -152,7 +155,7 @@ class _FlightDetailScreenState extends State<FlightDetailScreen> {
     try {
       final editedSite = await showDialog<Site>(
         context: context,
-        builder: (context) => _EditSiteDialog(site: _launchSite!),
+        builder: (context) => EditSiteDialog(site: _launchSite!),
       );
       
       if (editedSite != null && mounted) {
@@ -185,9 +188,9 @@ class _FlightDetailScreenState extends State<FlightDetailScreen> {
       
       if (!mounted) return;
       
-      final result = await showDialog<_WingSelectionResult>(
+      final result = await showDialog<WingSelectionResult>(
         context: context,
-        builder: (context) => _WingSelectionDialog(
+        builder: (context) => WingSelectionDialog(
           wings: wings,
           currentWing: _wing,
         ),
@@ -1340,911 +1343,3 @@ class _FlightDetailScreenState extends State<FlightDetailScreen> {
 
 }
 
-// Result wrapper to distinguish between cancellation and selection
-class _SiteSelectionResult {
-  final Site? selectedSite;
-  
-  const _SiteSelectionResult(this.selectedSite);
-}
-
-// Result wrapper for wing selection
-class _WingSelectionResult {
-  final Wing? selectedWing;
-  
-  const _WingSelectionResult(this.selectedWing);
-}
-
-// Custom dialog for site selection with search functionality
-class _SiteSelectionDialog extends StatefulWidget {
-  final List<Site> sites;
-  final Site? currentSite;
-  final String title;
-
-  const _SiteSelectionDialog({
-    required this.sites,
-    required this.currentSite,
-    required this.title,
-  });
-
-  @override
-  State<_SiteSelectionDialog> createState() => _SiteSelectionDialogState();
-}
-
-class _SiteSelectionDialogState extends State<_SiteSelectionDialog> {
-  late List<Site> _filteredSites;
-  final TextEditingController _searchController = TextEditingController();
-  Site? _selectedSite;
-  String _searchQuery = '';
-
-  @override
-  void initState() {
-    super.initState();
-    _filteredSites = widget.sites;
-    _selectedSite = widget.currentSite;
-    
-    // Sort sites alphabetically for easier browsing
-    _filteredSites.sort((a, b) => a.name.compareTo(b.name));
-  }
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  void _filterSites(String query) {
-    setState(() {
-      _searchQuery = query.toLowerCase();
-      if (_searchQuery.isEmpty) {
-        _filteredSites = List.from(widget.sites)
-          ..sort((a, b) => a.name.compareTo(b.name));
-      } else {
-        _filteredSites = widget.sites
-            .where((site) => 
-                site.name.toLowerCase().contains(_searchQuery) ||
-                (site.country?.toLowerCase().contains(_searchQuery) ?? false))
-            .toList()
-          ..sort((a, b) {
-            // Prioritize sites that start with the search query
-            final aStarts = a.name.toLowerCase().startsWith(_searchQuery);
-            final bStarts = b.name.toLowerCase().startsWith(_searchQuery);
-            if (aStarts && !bStarts) return -1;
-            if (!aStarts && bStarts) return 1;
-            return a.name.compareTo(b.name);
-          });
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // Group sites by country for better organization
-    final Map<String, List<Site>> countryGroupedSites = {};
-    for (final site in _filteredSites) {
-      final country = site.country ?? 'Unknown Country';
-      countryGroupedSites.putIfAbsent(country, () => []).add(site);
-    }
-    
-    // Sort countries alphabetically, but keep "Unknown Country" at the end
-    final sortedCountries = countryGroupedSites.keys.toList()..sort((a, b) {
-      if (a == 'Unknown Country' && b != 'Unknown Country') return 1;
-      if (a != 'Unknown Country' && b == 'Unknown Country') return -1;
-      return a.compareTo(b);
-    });
-    
-    // Sort sites within each country
-    for (final country in sortedCountries) {
-      countryGroupedSites[country]!.sort((a, b) => a.name.compareTo(b.name));
-    }
-
-    return AlertDialog(
-      title: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(widget.title),
-          const SizedBox(height: 16),
-          TextField(
-            controller: _searchController,
-            decoration: InputDecoration(
-              hintText: 'Search sites or countries...',
-              prefixIcon: const Icon(Icons.search),
-              border: const OutlineInputBorder(),
-              contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              suffixIcon: _searchQuery.isNotEmpty
-                  ? IconButton(
-                      icon: const Icon(Icons.clear),
-                      onPressed: () {
-                        _searchController.clear();
-                        _filterSites('');
-                      },
-                    )
-                  : null,
-            ),
-            onChanged: _filterSites,
-            autofocus: true,
-          ),
-          if (_filteredSites.isNotEmpty)
-            Padding(
-              padding: const EdgeInsets.only(top: 8),
-              child: Text(
-                '${_filteredSites.length} site${_filteredSites.length != 1 ? 's' : ''} found',
-                style: Theme.of(context).textTheme.bodySmall,
-              ),
-            ),
-        ],
-      ),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 400, // Fixed height for better UX
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // No site option
-            ListTile(
-              title: const Text('No site'),
-              leading: Radio<Site?>(
-                value: null,
-                groupValue: _selectedSite,
-                onChanged: (Site? value) {
-                  setState(() {
-                    _selectedSite = value;
-                  });
-                },
-              ),
-              onTap: () {
-                setState(() {
-                  _selectedSite = null;
-                });
-              },
-              dense: true,
-            ),
-            const Divider(),
-            // Site list
-            Expanded(
-              child: _filteredSites.isEmpty
-                  ? Center(
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Icon(Icons.search_off, 
-                               size: 48, 
-                               color: Colors.grey[400]),
-                          const SizedBox(height: 16),
-                          Text(
-                            'No sites match your search',
-                            style: TextStyle(color: Colors.grey[600]),
-                          ),
-                        ],
-                      ),
-                    )
-                  : ListView.builder(
-                      itemCount: _searchQuery.isEmpty 
-                          ? sortedCountries.fold<int>(0, (count, country) => 
-                              count + 1 + countryGroupedSites[country]!.length) // Country header + sites
-                          : _filteredSites.length,
-                      itemBuilder: (context, index) {
-                        if (_searchQuery.isNotEmpty) {
-                          // When searching, show flat list
-                          final site = _filteredSites[index];
-                          return ListTile(
-                            title: Text(
-                              site.name,
-                              style: site.name == 'Unknown'
-                                  ? TextStyle(fontStyle: FontStyle.italic,
-                                            color: Colors.grey[600])
-                                  : null,
-                            ),
-                            subtitle: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                if (site.altitude != null)
-                                  Text('${site.altitude!.toInt()} m'),
-                                if (site.country != null)
-                                  Text(
-                                    site.country!,
-                                    style: TextStyle(
-                                      fontSize: 12,
-                                      color: Colors.grey[600],
-                                    ),
-                                  ),
-                              ],
-                            ),
-                            leading: Radio<Site>(
-                              value: site,
-                              groupValue: _selectedSite,
-                              onChanged: (Site? value) {
-                                setState(() {
-                                  _selectedSite = value;
-                                });
-                              },
-                            ),
-                            onTap: () {
-                              setState(() {
-                                _selectedSite = site;
-                              });
-                            },
-                            dense: true,
-                          );
-                        } else {
-                          // When not searching, show hierarchical structure
-                          return _buildHierarchicalItem(index, countryGroupedSites);
-                        }
-                      },
-                    ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(), // Return null for cancellation
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          onPressed: () => Navigator.of(context).pop(_SiteSelectionResult(_selectedSite)),
-          child: const Text('Select'),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildHierarchicalItem(int index, Map<String, List<Site>> countryGroupedSites) {
-    // Calculate which item we're showing based on the hierarchical structure
-    int currentIndex = 0;
-    
-    for (final country in countryGroupedSites.keys.toList()..sort((a, b) {
-      if (a == 'Unknown Country' && b != 'Unknown Country') return 1;
-      if (a != 'Unknown Country' && b == 'Unknown Country') return -1;
-      return a.compareTo(b);
-    })) {
-      if (currentIndex == index) {
-        // Country header
-        return Container(
-          width: double.infinity,
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-          margin: const EdgeInsets.only(top: 8),
-          color: Theme.of(context).colorScheme.primary.withValues(alpha: 0.1),
-          child: Text(
-            country,
-            style: Theme.of(context).textTheme.titleMedium?.copyWith(
-              fontWeight: FontWeight.bold,
-              color: Theme.of(context).colorScheme.primary,
-            ),
-          ),
-        );
-      }
-      currentIndex++;
-      
-      final sites = countryGroupedSites[country]!;
-      for (final site in sites) {
-        if (currentIndex == index) {
-          // Site item
-          return Padding(
-            padding: const EdgeInsets.only(left: 16),
-            child: ListTile(
-              title: Text(
-                site.name,
-                style: site.name == 'Unknown'
-                    ? TextStyle(fontStyle: FontStyle.italic,
-                              color: Colors.grey[600])
-                    : null,
-              ),
-              subtitle: site.altitude != null
-                  ? Text('${site.altitude!.toInt()} m')
-                  : null,
-              leading: Radio<Site>(
-                value: site,
-                groupValue: _selectedSite,
-                onChanged: (Site? value) {
-                  setState(() {
-                    _selectedSite = value;
-                  });
-                },
-              ),
-              onTap: () {
-                setState(() {
-                  _selectedSite = site;
-                });
-              },
-              dense: true,
-            ),
-          );
-        }
-        currentIndex++;
-      }
-    }
-    
-    // Fallback (shouldn't reach here)
-    return const SizedBox.shrink();
-  }
-}
-
-// Custom wing selection dialog
-class _WingSelectionDialog extends StatefulWidget {
-  final List<Wing> wings;
-  final Wing? currentWing;
-
-  const _WingSelectionDialog({
-    required this.wings,
-    required this.currentWing,
-  });
-
-  @override
-  State<_WingSelectionDialog> createState() => _WingSelectionDialogState();
-}
-
-class _WingSelectionDialogState extends State<_WingSelectionDialog> {
-  late List<Wing> _filteredWings;
-  final TextEditingController _searchController = TextEditingController();
-  Wing? _selectedWing;
-  String _searchQuery = '';
-
-  @override
-  void initState() {
-    super.initState();
-    _filteredWings = widget.wings;
-    _selectedWing = widget.currentWing;
-    
-    // Sort wings by manufacturer and model for easier browsing
-    _filteredWings.sort((a, b) {
-      final aManufacturer = a.manufacturer ?? '';
-      final bManufacturer = b.manufacturer ?? '';
-      final manufacturerCompare = aManufacturer.compareTo(bManufacturer);
-      if (manufacturerCompare != 0) return manufacturerCompare;
-      
-      final aModel = a.model ?? '';
-      final bModel = b.model ?? '';
-      return aModel.compareTo(bModel);
-    });
-  }
-
-  @override
-  void dispose() {
-    _searchController.dispose();
-    super.dispose();
-  }
-
-  void _filterWings(String query) {
-    setState(() {
-      _searchQuery = query.toLowerCase();
-      if (_searchQuery.isEmpty) {
-        _filteredWings = List.from(widget.wings)
-          ..sort((a, b) {
-            final aManufacturer = a.manufacturer ?? '';
-            final bManufacturer = b.manufacturer ?? '';
-            final manufacturerCompare = aManufacturer.compareTo(bManufacturer);
-            if (manufacturerCompare != 0) return manufacturerCompare;
-            
-            final aModel = a.model ?? '';
-            final bModel = b.model ?? '';
-            return aModel.compareTo(bModel);
-          });
-      } else {
-        _filteredWings = widget.wings
-            .where((wing) => 
-                (wing.manufacturer?.toLowerCase().contains(_searchQuery) ?? false) ||
-                (wing.model?.toLowerCase().contains(_searchQuery) ?? false))
-            .toList()
-          ..sort((a, b) {
-            final aManufacturer = a.manufacturer ?? '';
-            final bManufacturer = b.manufacturer ?? '';
-            final manufacturerCompare = aManufacturer.compareTo(bManufacturer);
-            if (manufacturerCompare != 0) return manufacturerCompare;
-            
-            final aModel = a.model ?? '';
-            final bModel = b.model ?? '';
-            return aModel.compareTo(bModel);
-          });
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const Text('Select Wing'),
-          const SizedBox(height: 16),
-          TextField(
-            controller: _searchController,
-            decoration: InputDecoration(
-              hintText: 'Search manufacturer or model...',
-              prefixIcon: const Icon(Icons.search),
-              border: const OutlineInputBorder(),
-              contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              suffixIcon: _searchQuery.isNotEmpty
-                  ? IconButton(
-                      icon: const Icon(Icons.clear),
-                      onPressed: () {
-                        _searchController.clear();
-                        _filterWings('');
-                      },
-                    )
-                  : null,
-            ),
-            onChanged: _filterWings,
-            autofocus: true,
-          ),
-          if (_filteredWings.isNotEmpty)
-            Padding(
-              padding: const EdgeInsets.only(top: 8),
-              child: Text(
-                '${_filteredWings.length} wing${_filteredWings.length != 1 ? 's' : ''} found',
-                style: Theme.of(context).textTheme.bodySmall,
-              ),
-            ),
-        ],
-      ),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 400,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // No wing option
-            ListTile(
-              title: const Text('No wing'),
-              leading: Radio<Wing?>(
-                value: null,
-                groupValue: _selectedWing,
-                onChanged: (Wing? value) {
-                  setState(() {
-                    _selectedWing = value;
-                  });
-                },
-              ),
-              onTap: () {
-                setState(() {
-                  _selectedWing = null;
-                });
-              },
-              dense: true,
-            ),
-            const Divider(),
-            // Wing list
-            Expanded(
-              child: _filteredWings.isEmpty
-                  ? Center(
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Icon(Icons.search_off, 
-                               size: 48, 
-                               color: Colors.grey[400]),
-                          const SizedBox(height: 16),
-                          Text(
-                            'No wings match your search',
-                            style: TextStyle(color: Colors.grey[600]),
-                          ),
-                        ],
-                      ),
-                    )
-                  : ListView.builder(
-                      itemCount: _filteredWings.length,
-                      itemBuilder: (context, index) {
-                        final wing = _filteredWings[index];
-                        return ListTile(
-                          title: Text('${wing.manufacturer ?? 'Unknown'} ${wing.model ?? 'Unknown'}'),
-                          subtitle: wing.size != null
-                              ? Text('Size: ${wing.size}')
-                              : null,
-                          leading: Radio<Wing>(
-                            value: wing,
-                            groupValue: _selectedWing,
-                            onChanged: (Wing? value) {
-                              setState(() {
-                                _selectedWing = value;
-                              });
-                            },
-                          ),
-                          onTap: () {
-                            setState(() {
-                              _selectedWing = wing;
-                            });
-                          },
-                          dense: true,
-                        );
-                      },
-                    ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(),
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          onPressed: () => Navigator.of(context).pop(_WingSelectionResult(_selectedWing)),
-          child: const Text('Select'),
-        ),
-      ],
-    );
-  }
-}
-
-// Site edit dialog - reused from manage sites screen
-class _EditSiteDialog extends StatefulWidget {
-  final Site site;
-
-  const _EditSiteDialog({required this.site});
-
-  @override
-  State<_EditSiteDialog> createState() => _EditSiteDialogState();
-}
-
-class _EditSiteDialogState extends State<_EditSiteDialog> with SingleTickerProviderStateMixin {
-  late TextEditingController _nameController;
-  late TextEditingController _latitudeController;
-  late TextEditingController _longitudeController;
-  late TextEditingController _altitudeController;
-  late TextEditingController _countryController;
-  late TabController _tabController;
-  MapController? _mapController;
-  bool _showSatelliteView = false;
-  final _formKey = GlobalKey<FormState>();
-
-  @override
-  void initState() {
-    super.initState();
-    _tabController = TabController(length: 2, vsync: this);
-    _nameController = TextEditingController(text: widget.site.name);
-    _latitudeController = TextEditingController(text: widget.site.latitude.toString());
-    _longitudeController = TextEditingController(text: widget.site.longitude.toString());
-    _altitudeController = TextEditingController(
-      text: widget.site.altitude?.toString() ?? '',
-    );
-    _countryController = TextEditingController(
-      text: widget.site.country ?? '',
-    );
-  }
-
-  @override
-  void dispose() {
-    _tabController.dispose();
-    _nameController.dispose();
-    _latitudeController.dispose();
-    _longitudeController.dispose();
-    _altitudeController.dispose();
-    _countryController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      child: SizedBox(
-        width: MediaQuery.of(context).size.width * 0.9,
-        height: MediaQuery.of(context).size.height * 0.8,
-        child: Column(
-          children: [
-            // Header with title and close button
-            Container(
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                border: Border(
-                  bottom: BorderSide(color: Theme.of(context).dividerColor),
-                ),
-              ),
-              child: Row(
-                children: [
-                  const Text(
-                    'Edit Site',
-                    style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-                  ),
-                  const Spacer(),
-                  IconButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    icon: const Icon(Icons.close),
-                  ),
-                ],
-              ),
-            ),
-            // Tab bar
-            TabBar(
-              controller: _tabController,
-              tabs: const [
-                Tab(
-                  icon: Icon(Icons.edit),
-                  text: 'Details',
-                ),
-                Tab(
-                  icon: Icon(Icons.map),
-                  text: 'Location',
-                ),
-              ],
-            ),
-            // Tab content
-            Expanded(
-              child: TabBarView(
-                controller: _tabController,
-                children: [
-                  _buildDetailsTab(),
-                  _buildLocationTab(),
-                ],
-              ),
-            ),
-            // Action buttons
-            Container(
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                border: Border(
-                  top: BorderSide(color: Theme.of(context).dividerColor),
-                ),
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  TextButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    child: const Text('Cancel'),
-                  ),
-                  const SizedBox(width: 8),
-                  FilledButton(
-                    onPressed: _saveChanges,
-                    child: const Text('Save'),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildDetailsTab() {
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Form(
-        key: _formKey,
-        child: SingleChildScrollView(
-          child: Column(
-            children: [
-              TextFormField(
-                controller: _nameController,
-                decoration: const InputDecoration(
-                  labelText: 'Site Name',
-                  border: OutlineInputBorder(),
-                ),
-                validator: (value) {
-                  if (value == null || value.trim().isEmpty) {
-                    return 'Site name is required';
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 16),
-              Row(
-                children: [
-                  Expanded(
-                    child: TextFormField(
-                      controller: _latitudeController,
-                      decoration: const InputDecoration(
-                        labelText: 'Latitude',
-                        border: OutlineInputBorder(),
-                      ),
-                      keyboardType: const TextInputType.numberWithOptions(decimal: true),
-                      validator: (value) {
-                        if (value == null || value.trim().isEmpty) {
-                          return 'Required';
-                        }
-                        final lat = double.tryParse(value.trim());
-                        if (lat == null || lat < -90 || lat > 90) {
-                          return 'Invalid latitude';
-                        }
-                        return null;
-                      },
-                    ),
-                  ),
-                  const SizedBox(width: 16),
-                  Expanded(
-                    child: TextFormField(
-                      controller: _longitudeController,
-                      decoration: const InputDecoration(
-                        labelText: 'Longitude',
-                        border: OutlineInputBorder(),
-                      ),
-                      keyboardType: const TextInputType.numberWithOptions(decimal: true),
-                      validator: (value) {
-                        if (value == null || value.trim().isEmpty) {
-                          return 'Required';
-                        }
-                        final lon = double.tryParse(value.trim());
-                        if (lon == null || lon < -180 || lon > 180) {
-                          return 'Invalid longitude';
-                        }
-                        return null;
-                      },
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              TextFormField(
-                controller: _altitudeController,
-                decoration: const InputDecoration(
-                  labelText: 'Altitude (m)',
-                  border: OutlineInputBorder(),
-                  hintText: 'Optional',
-                ),
-                keyboardType: const TextInputType.numberWithOptions(decimal: true),
-                validator: (value) {
-                  if (value != null && value.trim().isNotEmpty) {
-                    final alt = double.tryParse(value.trim());
-                    if (alt == null) {
-                      return 'Invalid altitude';
-                    }
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 16),
-              TextFormField(
-                controller: _countryController,
-                decoration: const InputDecoration(
-                  labelText: 'Country',
-                  border: OutlineInputBorder(),
-                  hintText: 'Optional',
-                ),
-                validator: (value) {
-                  // Country is optional, no validation needed
-                  return null;
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildLocationTab() {
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        children: [
-          Text(
-            widget.site.name,
-            style: Theme.of(context).textTheme.titleMedium,
-          ),
-          const SizedBox(height: 8),
-          Text(
-            '${widget.site.latitude.toStringAsFixed(6)}, ${widget.site.longitude.toStringAsFixed(6)}',
-            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-              fontFamily: 'monospace',
-            ),
-          ),
-          const SizedBox(height: 16),
-          Expanded(
-            child: _buildLocationMap(),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _saveChanges() {
-    if (_formKey.currentState!.validate()) {
-      final updatedSite = widget.site.copyWith(
-        name: _nameController.text.trim(),
-        latitude: double.parse(_latitudeController.text.trim()),
-        longitude: double.parse(_longitudeController.text.trim()),
-        altitude: _altitudeController.text.trim().isEmpty
-            ? null
-            : double.parse(_altitudeController.text.trim()),
-        country: _countryController.text.trim().isEmpty
-            ? null
-            : _countryController.text.trim(),
-        customName: true, // Mark as custom since user edited it
-      );
-      Navigator.of(context).pop(updatedSite);
-    }
-  }
-
-  void _toggleSatelliteView() {
-    setState(() {
-      _showSatelliteView = !_showSatelliteView;
-    });
-  }
-
-  Widget _buildLocationMap() {
-    _mapController ??= MapController();
-    
-    return Container(
-      decoration: BoxDecoration(
-        border: Border.all(color: Theme.of(context).dividerColor),
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Stack(
-        children: [
-          ClipRRect(
-            borderRadius: BorderRadius.circular(8),
-            child: FlutterMap(
-              mapController: _mapController,
-              options: MapOptions(
-                initialCenter: LatLng(widget.site.latitude, widget.site.longitude),
-                initialZoom: 14.0,
-                minZoom: 5.0,
-                maxZoom: 18.0,
-              ),
-              children: [
-                TileLayer(
-                  urlTemplate: _showSatelliteView 
-                    ? 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}'
-                    : 'https://tile.openstreetmap.org/{z}/{x}/{y}.png',
-                  userAgentPackageName: 'com.example.free_flight_log_app',
-                ),
-                MarkerLayer(
-                  markers: [
-                    Marker(
-                      point: LatLng(widget.site.latitude, widget.site.longitude),
-                      width: 40,
-                      height: 40,
-                      child: const Icon(
-                        Icons.location_pin,
-                        color: Colors.red,
-                        size: 40,
-                      ),
-                    ),
-                  ],
-                ),
-                // Attribution overlay for satellite tiles
-                if (_showSatelliteView)
-                  Align(
-                    alignment: Alignment.bottomRight,
-                    child: Container(
-                      margin: const EdgeInsets.all(4),
-                      padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 2),
-                      decoration: BoxDecoration(
-                        color: Colors.white.withValues(alpha: 0.8),
-                        borderRadius: BorderRadius.circular(2),
-                      ),
-                      child: Text(
-                        'Powered by Esri',
-                        style: TextStyle(
-                          fontSize: 10,
-                          color: Colors.black87,
-                        ),
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-          ),
-          // Satellite toggle button
-          Positioned(
-            top: 8,
-            right: 8,
-            child: Container(
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.circular(4),
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.black26,
-                    blurRadius: 4,
-                    offset: Offset(0, 2),
-                  ),
-                ],
-              ),
-              child: IconButton(
-                onPressed: _toggleSatelliteView,
-                icon: Icon(
-                  _showSatelliteView ? Icons.map : Icons.satellite_alt,
-                  size: 20,
-                ),
-                tooltip: _showSatelliteView ? 'Street View' : 'Satellite View',
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
