--- free_flight_log_app/lib/presentation/screens/simple_flight_list_screen.dart
@@ -1,53 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SimpleFlightListScreen extends StatelessWidget {
-  const SimpleFlightListScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Free Flight Log'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.flight_takeoff,
-              size: 64,
-              color: Colors.grey[400],
-            ),
-            const SizedBox(height: 16),
-            Text(
-              'Free Flight Log App',
-              style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                    color: Colors.grey[600],
-                  ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              'Database and UI ready!',
-              style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Colors.grey[500],
-                  ),
-            ),
-            const SizedBox(height: 32),
-            ElevatedButton.icon(
-              onPressed: () {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('Flight form coming soon!'),
-                  ),
-                );
-              },
-              icon: const Icon(Icons.add),
-              label: const Text('Add Flight'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
