+++ free_flight_log_app/test/site_matching_test.dart
@@ -15,9 +15,9 @@ void main() {
       expect(siteMatchingService.siteCount, greaterThan(0));
     });
 
-    test('Should find Interlaken site near Beatenberg coordinates', () {
+    test('Should find Interlaken site near Beatenberg coordinates', () async {
       // Coordinates near Interlaken - Beatenberg (from sample data)
-      final site = siteMatchingService.findNearestLaunchSite(46.6945, 7.9867);
+      final site = await siteMatchingService.findNearestLaunchSite(46.6945, 7.9867);
       
       expect(site, isNotNull);
       expect(site!.name, equals('Interlaken - Beatenberg'));
@@ -25,18 +25,18 @@ void main() {
       expect(site.siteType, equals('launch'));
     });
 
-    test('Should find landing site near Interlaken', () {
+    test('Should find landing site near Interlaken', () async {
       // Coordinates near Interlaken Landing Field
-      final site = siteMatchingService.findNearestLandingSite(46.6774, 7.8636);
+      final site = await siteMatchingService.findNearestLandingSite(46.6774, 7.8636);
       
       expect(site, isNotNull);
       expect(site!.name, equals('Interlaken Landing Field'));
       expect(site.siteType, equals('landing'));
     });
 
-    test('Should return null for coordinates far from any site', () {
+    test('Should return null for coordinates far from any site', () async {
       // Coordinates in the middle of the ocean
-      final site = siteMatchingService.findNearestSite(0.0, 0.0, maxDistance: 100);
+      final site = await siteMatchingService.findNearestSite(0.0, 0.0, maxDistance: 100);
       
       expect(site, isNull);
     });
@@ -48,9 +48,9 @@ void main() {
       expect(results.first.name.toLowerCase(), contains('interlaken'));
     });
 
-    test('Should get site name suggestion with coordinates fallback', () {
+    test('Should get site name suggestion with coordinates fallback', () async {
       // Test with known site coordinates
-      final knownSiteName = siteMatchingService.getSiteNameSuggestion(
+      final knownSiteName = await siteMatchingService.getSiteNameSuggestion(
         46.6945, 7.9867,
         prefix: 'Launch',
         siteType: 'launch',
@@ -58,7 +58,7 @@ void main() {
       expect(knownSiteName, equals('Launch Interlaken - Beatenberg'));
       
       // Test with unknown coordinates (should fall back to coordinates)
-      final unknownSiteName = siteMatchingService.getSiteNameSuggestion(
+      final unknownSiteName = await siteMatchingService.getSiteNameSuggestion(
         0.0, 0.0,
         prefix: 'Launch',
         siteType: 'launch',
